backend/requirements.txt
numpy>=1.15.0
pandas
scipy
keras
tensorflow
google-cloud-storage